- add test cases - move PSE related into media-center.js - cleanup
